+++ backend/src/main/java/hanglog/community/domain/repository/LikeRepository.java
@@ -1,6 +1,6 @@
 package hanglog.community.domain.repository;
 
-import hanglog.community.domain.LikeInfoDto;
+import hanglog.community.dto.LikeInfoByTrip;
 import hanglog.community.domain.Likes;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -16,11 +16,11 @@ public interface LikeRepository extends JpaRepository<Likes, Long> {
     Long countLikesByTripId(final Long tripId);
 
     @Query("""
-            SELECT new hanglog.community.domain.LikeInfoDto
+            SELECT new hanglog.community.dto.LikeInfoByTrip
             (l.tripId, COUNT(l.memberId), EXISTS(SELECT 1 FROM Likes l_1 WHERE l_1.memberId = :memberId AND l_1.tripId = l.tripId))
             FROM Likes l
             WHERE l.tripId in :tripIds
             GROUP BY l.tripId
              """)
-    List<LikeInfoDto> countByMemberIdAndTripId(@Param("memberId") final Long memberId, @Param("tripIds") final List<Long> tripIds);
+    List<LikeInfoByTrip> findLikeCountAndIsLikeByTripIds(@Param("memberId") final Long memberId, @Param("tripIds") final List<Long> tripIds);
 }
