+++ backend/src/main/java/hanglog/community/dto/LikeInfo.java
@@ -1,16 +1,12 @@
-package hanglog.community.domain;
+package hanglog.community.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.ToString;
 
 @Getter
 @AllArgsConstructor
-@ToString
 public class LikeInfo {
 
     private final long likeCount;
     private final boolean isLike;
-
-
 }
