+++ backend/src/main/java/hanglog/trip/domain/repository/TripCityRepository.java
@@ -1,6 +1,6 @@
 package hanglog.trip.domain.repository;
 
-import hanglog.community.domain.TripCityInfoDto;
+import hanglog.community.dto.CityInfoByTrip;
 import hanglog.trip.domain.TripCity;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -14,8 +14,8 @@ public interface TripCityRepository extends JpaRepository<TripCity, Long> {
     void deleteAllByTripId(Long tripId);
 
     @Query("""
-            SELECT new hanglog.community.domain.TripCityInfoDto
+            SELECT new hanglog.community.dto.CityInfoByTrip
             (tc.trip.id, tc.city) FROM TripCity tc WHERE tc.trip.id IN :tripIds
             """)
-    List<TripCityInfoDto> findTripIdAndCitiesByTripIds(@Param("tripIds") final List<Long> tripIds);
+    List<CityInfoByTrip> findTripIdAndCitiesByTripIds(@Param("tripIds") final List<Long> tripIds);
 }
